Update terms.json to match 18f/crime-data-api
We have been keeping an updated list of terms and definitions
in the 18f/crime-data-api repo. We should use that for the
<Glossary /> component, so I wrote a small script to pull the
YML file and convert it into a JSON file.

I think we can run this script during development to keep the
terms list up to date. When we have a better understanding of
how 18f/crime-data-style will function, we can probably move
the canonical version of this there to power the glossary
component