start to incorporate 18f/glossary (#21)
* start to incorporate 18f/glossary

* linting

* fix broken test

* lint

* Update terms.json to match 18f/crime-data-api

We have been keeping an updated list of terms and definitions
in the 18f/crime-data-api repo. We should use that for the
<Glossary /> component, so I wrote a small script to pull the
YML file and convert it into a JSON file.

I think we can run this script during development to keep the
terms list up to date. When we have a better understanding of
how 18f/crime-data-style will function, we can probably move
the canonical version of this there to power the glossary
component

* humble beginnings...

* add x and y axis components

* add circles for points

* add basic tests

* Add contributing document (#18)

* Add contributing document

Add a contribution policy as is regular practice for 18F repos.
It is not only for public contributors, but should also serve
as a reference document for 18F employees working on this project.

The "Experience principles" section needs to be reviewed and
further developed.

* Add basic install instructions to README

* Update contributing doc

Add guidance regarding branch naming for future reference

* Configure lint for devDependencies

The linter was throwing an error because it was configured
to show an error when a package was being imported from
devDependencies. As most of our front end packages are currently
listed as such (to build on CI and not cloud.gov) this was
causing a problem.

* Update README.md

Fix API link in README.

Committing straight to master since it is a tiny change.