Add <Breadcrumbs />
I am not sure if we'll need breadcrumbs outside of the "Explorer"
views, but for the time being I made a component to encapsulate styling
and behavior.

The component takes all the props from its parent component s that it
has access to all of the react-router information and can build links
correctly.