Add proper handler to <Term />
The Term component now expects a dispatch prop (a function) so that it
can dispatch the appropriate action to show the term.

Note that the ID prop should include spaces as it is used to query for
the term directly. Not sure if it would be better to use a different
name, but having a "term" prop on the <Term /> component felt redundant.