Add location helpers to history util
The history.push method expects react-router location objects, which
are structured in a particular way. Added a helper function to take a
current location object and return a new one that reflects the change.

Added to the history utility since it is used where history.push would
be and therefore is easy to import alongside