Include population data for now to calculate rates on the fly
We want to show rates to control crime for population, but the API does
not yet supply that on a yearly basis
(https://github.com/18F/crime-data-api/issues/290).

Right now we are using the most recent year's population from the
Census and using it to calculate the crime rate for each year in a
query. This is absolutely not a good practice, but we're doing it for
user testing tomorrow.