Rename incidents to nibrs
I changed the names of things In places where we use "incidents" to
"nibrs" or "nibrsDimension". I did this because it is more explicit
once you are familiar with the data model.

Incidents felt wrong because we're actually  just using it for NIBRS
dimensions and aggregated counts. I thought that at some point we
actually want might an incidents reducer and display them in the UI