Use NIBRS API data (#182)
* Fetch NIBRS dimension data from the API

* Start to use API data in Explorer view

* Rename incidents to nibrs

I changed the names of things In places where we use "incidents" to
"nibrs" or "nibrsDimension". I did this because it is more explicit
once you are familiar with the data model.

Incidents felt wrong because we're actually  just using it for NIBRS
dimensions and aggregated counts. I thought that at some point we
actually want might an incidents reducer and display them in the UI

* Change state codes

https://gsa-tts.slack.com/archives/fbi-cde/p1485791870002035

* Use some of the API data in NIBRS container

* Use <Loading /> as a placeholder component

* Refactor some trend chart stuff in <TrendContainer />

* Remove api test for now

* Adjust summary endpoint

* Fix footer glossary link

* Fix code climate complaints

* Filter nibrs related data to only include year range