Show errors for time period filter (#197)
* Show errors for time period filter

* Fix code climate errors

* Change filter title to location from placename

* Move error below fields

* Use props instead of state for <TimePeriodFilter />

https://github.com/18F/crime-data-explorer/pull/197#discussion_r98957336

* Don't scroll to top on route changes

Because the sidebar stretches all the way down the page, its pretty
annoying if you have to scroll down after selecting each year, one by
one. Now, the content updates but the page doesn't jump to the top.

This should work just fine for the links available in the header since
the user will already be scrolled to the top of the page then. I added
a scroll function to the footer links so that the page is scrolled to
the top for those.