Add colors prop to <UsaMap />
The map component is great but we want to be able to change the colors
of the states based on where the component is shown. To this end, I
added a colors prop which is a mapping of state IDs to css classes.

If this prop is not supplied (such as on the home page), the default
styling is used.