Use different state ID mapping for bulk nibrs download
Sometimes we use abbreviations when talking to the API and sometimes we
need to use state IDs. There are a few different numerical state IDs
that can get assigned, so I moved them into the components that
actually use them to reduce confusion about where things are used.

I also updated the IDs to use the ones supplied in
https://github.com/18F/crime-data-explorer/issues/410#issuecomment-28279
7779

Fixes #411