Change data download names (#537)
* Add propTypes to <DownloadDataBtn />

* Move data download button to be in <NibrsCard />

The download button should give the user data from the entire card, not
just the last table component. I moved the data download button out and
into the NibrsCard, passing in the crime and place name so the
downloaded file has a useful name.

Right now the data download button only returns the first item in the
data array, but it should actually trigger multiple downloads if there
are multiple datasets inside of the nibrs card.