Server side rendering (#521)
* trying stuff

* Make footer links behave the same as header

* Remove <Glossary /> component from app

Temporarily removing glossary component to deal with server side
rendering

* Refactor server.js to use new ES features

* add redux to server2, webpack server watch

* Refactor build steps for server side rendering

To get the app to be rendered server side we need some changes to our
build steps. In addition, we want to reuse as much code as possible so
I changed the location the app uses for built assets. Also use a
browser history to remove the hash

* Clear local storage routes for easier debugging

* Request data from root for <StateThumbnail />

* Check for window object when using

Now that the code will be executed in an environment where window might
not exist, we should check to make sure it is available before trying
to use it.

* Create store in server and apply middelware

* Run on cloud foundry instances

Make sure all dependencies need to run the app on cloud foundry, but
not to build it, are in the dependencies

* Fix some linting issues with server webpack config

* Disable some linting on server.js and Glossary.js

* Use old eslint on codeclimate

* Remove comments and unused code from server.js

* Simplify build steps

https://github.com/18F/crime-data-explorer/pull/521#discussion_r10496616
4

* Move static assets to public directory

Use the public directory to serve specific static assets such as images
or robots.txt