Fetch national participation data
Now that we fetch the national participation data from the API, we
cannot assume it will just be there when we need to draw trend lines.

Because shouldFetchUcr will return true for the nationalKey when it is
the place and we don't want to fire off two requests for the same data,
fetchNational will be false when the place is national.