Use env var to set API host
To faciliate using two different API servers, depending on if
the app is a tagged release or not (and therefore using the demo.yml
manifest), I refactored src/server.js to source the API host from
the CDE_API environment variable.

I updated the manifest files to include this new environment
variable and set it with the appropriate API host.