Gzip bundle.js in build process
To reduce the size of the file we transfer over the network,
we want to gzip, at least, the javascript bundle.js file.

Using some middleware that checks for a compressed file (by
adding .gz to the requested file name) and serves it if
available. Otherwise, serves the static files just like
before.