Zip CSVs to trigger a single download
The browser alert/permission request to download multiple files is
subtle on many popular browsers so it was confusing people as to why
their files didn't download. Instead, we can wrap all the files into a
single .zip file so that the user can easily get all the relevant data.