Add /feedback endpoint for creating github issues
Checks for the presence of the necessary configuration variables and
checks (if in production) to make sure that the request is coming from
one of our URLs. This wouldn't prevent somebody from making a request
in the console, but would does prevent other clients from being able to
create issues in our repos.