Only make agencies with some data active in search results
If either months_reported or nibrs_months_reported are not equal to
zero, then show the agency with black text and respond to click events
in the search results. If that condition is not met, then show them as
gray and do not respond to click events for that item.

Refs #913