Summary reducer should maintain existing data
The summary reducer shouldn't write over existing data that isn't
relevant to the current action. Instead, it should continue to increase
the summary object so that it reflects all the data gathered from the
API. This will enable us to do some smarter checking of what exists in
the state before we make fetch requests.

Currently, we lean heavily on the service worker to account for making
all of these requests over and over again. However, service workers
aren't yet available in many browsers or platforms. Therefore, we
should also do smart checking before dispatching API requests to reduce
the network traffic initiated by the application and increase the
responsiveness.