Use "offense" instead of "incident" in the NIBRS intro text
We are technically counting the number of reported offenses, not the
number of incidents which has a specific meaning in NIBRS. Incidents
can have many offenses associated with them, and most closely
correlates with a row of data. However, when regular folks are reading
this, the distinction is lost. So we'll use "offenses" since it is
technically more correct and has a very similar meaning to folks not
familiar with the NIBRS definition.