Only show two trend lines (#1183)
* Restrict summaries passed to trend container

Since summaries are never removed, we have to be selective about which
are passed from the state to the trendcontainer component. We should
only include the national data and the state, if applicable.

Include some file mocking so that the content util can be properly
imported.

* Show selected trend line in red

If the national view is selected, then the single trend line should be
red. If a state is selected, then the state should be red.

* Make Home component a container and give prop-types

* Pass reliable places to TrendContainer instead of changing data

Instead of changing the summary data that is passed to the container,
let's just pass an array of places that can be used reliably inside of
the component per @brendansudol's recommendation.

https://github.com/18F/crime-data-frontend/pull/1183#discussion_r1298726
94