Use bindActionCreators for page level components
Removing the dispatch prop in favor of an object called actions that is
the necessary action creators bound with the dispatch function by the
time they are passed into the component.

Also reduce the appState prop that is passed in, particularly to
<Explorer /> so that page level components can use connect() to get the
parts of app state that they care about and it isn't passed in from
higher props.