Refactor mungeSummaryData util function
The mungeSummary data function was doing quite a bit all in one call
which made it require a lot of parameters and was harder to follow. It
has been refactored into a series of smaller helper functions that are
used at different points in the app. This way the data is reduced into
a more usable form that requires no further calculation, just shape
manipulation and filtering.

Added tests for all new functions in summary util