Refactor trend chart to use simpler object
The createSeries function was generating an object that had dates,
rates and series along with `segments` and `gaps`. The latter two were
used when we were plotting trends of actual counts to handle the cases
where there were missing years in the data. However, now that the trend
charts use estimated data, there are no holes so we can remove all of
that logic.