Include a content security policy in the application (#1294)
Right now, this CSP is pretty wide open. This is because the current
implementation of the glossary tool relies on eval() and Google
Analytics relies on loading scripts and images.

However, this CSP should prevent other third party scripts from
executing. Once the glossary tool removes its dependency on the
templating library it uses, we can remove the inline permissions from
our CSP.

The easiest way to make sure the CSP works is to avoid inlining as much
as possible. Thus, I moved the script that powers the API docs to its
own file and allowed all the hosts that are required for analytics
tracking.